+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBSystemOperations.java
@@ -22,7 +22,9 @@ package com.apple.foundationdb.record.provider.foundationdb;
 
 import com.apple.foundationdb.Transaction;
 import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.record.logging.LogMessageKeys;
 import com.apple.foundationdb.system.SystemKeyspace;
+import com.google.common.collect.Lists;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -67,7 +69,7 @@ public class FDBSystemOperations {
             final Transaction tr = context.ensureActive();
             tr.options().setReadSystemKeys();
             return tr.get(SystemKeyspace.PRIMARY_DATACENTER_KEY).thenApply(FDBSystemOperations::nullableUtf8);
-        });
+        }, Lists.newArrayList(LogMessageKeys.TRANSACTION_NAME, "FDBSystemOperations::getPrimaryDatacenterAsync"));
     }
 
     /**
@@ -107,7 +109,8 @@ public class FDBSystemOperations {
      */
     @Nonnull
     public static CompletableFuture<String> getConnectionStringAsync(@Nonnull FDBDatabaseRunner runner) {
-        return runner.runAsync(FDBSystemOperations::getConnectionStringAsyncInternal);
+        return runner.runAsync(FDBSystemOperations::getConnectionStringAsyncInternal,
+                Lists.newArrayList(LogMessageKeys.TRANSACTION_NAME, "FDBSystemOperations::getConnectionStringAsync"));
     }
 
     /**
@@ -140,7 +143,8 @@ public class FDBSystemOperations {
      */
     @Nonnull
     public static CompletableFuture<String> getClusterFilePathAsync(@Nonnull FDBDatabaseRunner runner) {
-        return runner.runAsync(FDBSystemOperations::getClusterFilePathAsyncInternal);
+        return runner.runAsync(FDBSystemOperations::getClusterFilePathAsyncInternal,
+                Lists.newArrayList(LogMessageKeys.TRANSACTION_NAME, "FDBSystemOperations::getClusterFilePathAsync"));
     }
 
     /**