+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexer.java
@@ -45,6 +45,7 @@ import com.apple.foundationdb.subspace.Subspace;
 import com.apple.foundationdb.synchronizedsession.SynchronizedSession;
 import com.apple.foundationdb.tuple.Tuple;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Message;
 import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
@@ -612,7 +613,8 @@ public class OnlineIndexer implements AutoCloseable {
      */
     public CompletableFuture<Void> stopOngoingOnlineIndexBuildsAsync() {
         return runner.runAsync(context -> openRecordStore(context).thenAccept(recordStore ->
-                stopOngoingOnlineIndexBuilds(recordStore, index)));
+                stopOngoingOnlineIndexBuilds(recordStore, index)),
+                runAsyncDetails("OnlineIndexer::stopOngoingOnlineIndexBuilds"));
     }
 
     /**
@@ -647,7 +649,7 @@ public class OnlineIndexer implements AutoCloseable {
      */
     public CompletableFuture<Boolean> checkAnyOngoingOnlineIndexBuildsAsync() {
         return runner.runAsync(context -> openRecordStore(context).thenCompose(recordStore ->
-                checkAnyOngoingOnlineIndexBuildsAsync(recordStore, index)));
+                checkAnyOngoingOnlineIndexBuildsAsync(recordStore, index)), runAsyncDetails("OnlineIndexer::checkAnyOngoingOnlineIndexBuilds"));
     }
 
     /**
@@ -759,7 +761,7 @@ public class OnlineIndexer implements AutoCloseable {
                                     .thenApply(vignore2 -> true);
                         }
                     });
-        }));
+        }), runAsyncDetails("OnlineIndexer::markReadableIfBuilt"));
     }
 
     /**
@@ -772,7 +774,7 @@ public class OnlineIndexer implements AutoCloseable {
     @Nonnull
     public CompletableFuture<Boolean> markReadable() {
         return getRunner().runAsync(context -> openRecordStore(context)
-                .thenCompose(store -> store.markIndexReadable(index)));
+                .thenCompose(store -> store.markIndexReadable(index)), runAsyncDetails("OnlineIndexer::markReadable"));
     }
 
     /**
@@ -1988,6 +1990,14 @@ public class OnlineIndexer implements AutoCloseable {
     @Nonnull
     public static OnlineIndexer forRecordStoreAndIndex(@Nonnull FDBRecordStore recordStore, @Nonnull String index) {
         return newBuilder().setRecordStore(recordStore).setIndex(index).build();
+    }
+
+    protected List<Object> runAsyncDetails(String transactionName, Object ...keysAndValues) {
+        return new ImmutableList.Builder<>()
+                .add(LogMessageKeys.TRANSACTION_NAME, transactionName)
+                .add(common.indexLogMessageKeyValues())
+                .add(keysAndValues)
+                .build();
     }
 
     /**